fixes tests with multiple setting files
