+++ client/src/test/index.ts
@@ -18,44 +18,17 @@
 // BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT 
 // OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-import { assert } from 'console';
-import * as fs from 'fs';
 import * as glob from 'glob';
 import * as Mocha from 'mocha';
 import * as path from 'path';
 
-const PROJECT_ROOT = path.join(__dirname, "..", "..");
 const TESTS_ROOT = __dirname;
-const DATA_ROOT = path.join(PROJECT_ROOT, "src", "test", "data");
 
+/** 
+ * executes all test suites in the current workspace. It is assumed that 
+ * the workspace has been prepared (e.g. by placing a certain settings file)
+ */
 export async function run(): Promise<void> {
-    // iterate over different settings and execute the test suite for each
-    // settings file.
-
-    console.info("Reading list of settings...");
-	const settings_list = fs.readdirSync(path.join(DATA_ROOT, "settings")).sort();
-	assert(settings_list.length > 0, "There are no settings to test");
-
-    let firstIteration: boolean = true;
-    for (const settings_file of settings_list) {
-		console.info(`Testing with settings '${settings_file}'...`);
-        // first settings file gets copied by `runTest`
-        if (!firstIteration) {
-            // shutting down and activating the extension is done by TestHelper.ts
-            const settings_path = path.join(DATA_ROOT, "settings", settings_file);
-            const workspace_path = process.env["VIPER_IDE_WORKSPACE_PATH"];
-            const workspace_vscode_path = path.join(workspace_path, ".vscode");
-	        const workspace_settings_path = path.join(workspace_vscode_path, "settings.json");
-            fs.copyFileSync(settings_path, workspace_settings_path);
-            console.log(`settings file copied from ${settings_path} to ${workspace_settings_path}`);
-        }
-        firstIteration = false;
-
-        await runTests();
-	}
-}
-
-async function runTests(): Promise<void> {
     // Create the mocha test
 	const mocha = new Mocha({
 		ui: 'tdd',
