+++ client/src/test/runTest.ts
@@ -27,7 +27,12 @@ import { runTests } from '@vscode/test-electron';
 const PROJECT_ROOT = path.join(__dirname, "..", "..");
 const DATA_ROOT = path.join(PROJECT_ROOT, "src", "test", "data");
 
-async function main() {
+// the test setup is as follows:
+// there is a visual studio code instance (i.e. execute `runTests`) for each
+// settings file. All test suites for a given configuration (i.e. settings file)
+// are executed within the same visual studio code instance.
+
+async function main(): Promise<void> {
 	// Download VS Code, unzip it and run the integration test
 	console.info("Reading VS Code version...");
 	const vscode_version = fs.readFileSync(path.join(DATA_ROOT, "vscode-version")).toString().trim();
@@ -36,48 +41,60 @@ async function main() {
 	const settings_list = fs.readdirSync(path.join(DATA_ROOT, "settings")).sort();
 	assert(settings_list.length > 0, "There are no settings to test");
 
-    // The folder containing the Extension Manifest package.json
-	// Passed to `--extensionDevelopmentPath`
-	const extensionDevelopmentPath = PROJECT_ROOT;
+	let firstIteration = true;
+	for (const settings_file of settings_list) {
+		if (!firstIteration) {
+			// workaround for a weird "exit code 55" error that happens on
+			// macOS when starting a new vscode instance immediately after
+			// closing an old one. (by fpoli)
+			await new Promise(resolve => setTimeout(resolve, 5000));
+		}
+		firstIteration = false;
+		
+		console.info(`Testing with settings '${settings_file}'...`);
+		const settings_path = path.join(DATA_ROOT, "settings", settings_file);
+		
+		// The folder containing the Extension Manifest package.json
+		// Passed to `--extensionDevelopmentPath`
+		const extensionDevelopmentPath = PROJECT_ROOT;
 
-	// The path to the extension test script
-	// Passed to --extensionTestsPath
-	const extensionTestsPath = path.resolve(__dirname, 'index');
+		// The path to the extension test script
+		// Passed to --extensionTestsPath
+		const extensionTestsPath = path.resolve(__dirname, 'index');
 
-	const tmpWorkspace = tmp.dirSync({ unsafeCleanup: true });
-
-    // copy first settings file to tmpWorkspace:
-    const settings_path = path.join(DATA_ROOT, "settings", settings_list[0]);
-    const workspace_vscode_path = path.join(tmpWorkspace.name, ".vscode");
-	const workspace_settings_path = path.join(workspace_vscode_path, "settings.json");
-	fs.mkdirSync(workspace_vscode_path);
-	fs.copyFileSync(settings_path, workspace_settings_path);
-
-    // get environment variables
-	const env: NodeJS.ProcessEnv = process.env;
-	// add additional environment variables to
-	// - path to workspace folder (only used by `index.ts`)
-	// - auto accept confirmation messages of Viper-IDE
-	// - wipe global storage path to force install Viper Tools after each activation
-	env.VIPER_IDE_WORKSPACE_PATH = tmpWorkspace.name;
-	env.VIPER_IDE_ASSUME_YES = "1";
-	env.VIPER_IDE_CLEAN_INSTALL = "1";
-
-    try {
-        await runTests({
-			version: vscode_version,
-			extensionDevelopmentPath,
-			extensionTestsPath,
-			// note that passing environment variables seems to only work when invoking the tests via CLI
-			extensionTestsEnv: env,
-			// Disable any other extension
-			launchArgs: ["--disable-extensions", tmpWorkspace.name],
-		});
-    } finally {
+		const tmpWorkspace = tmp.dirSync({ unsafeCleanup: true });
 		try {
-			tmpWorkspace.removeCallback();
-		} catch (e) {
-			console.warn(`cleaning temporary directory has failed with error ${e}`);
+			// prepare workspace with the settings:
+			const workspace_vscode_path = path.join(tmpWorkspace.name, ".vscode");
+			const workspace_settings_path = path.join(workspace_vscode_path, "settings.json");
+			fs.mkdirSync(workspace_vscode_path);
+			fs.copyFileSync(settings_path, workspace_settings_path);
+
+			// get environment variables
+			const env: NodeJS.ProcessEnv = process.env;
+			// add additional environment variables to
+			// - path to workspace folder (only used by `index.ts`)
+			// - auto accept confirmation messages of Viper-IDE
+			// - wipe global storage path to force install Viper Tools after each activation
+			env.VIPER_IDE_WORKSPACE_PATH = tmpWorkspace.name;
+			env.VIPER_IDE_ASSUME_YES = "1";
+			env.VIPER_IDE_CLEAN_INSTALL = "1";
+
+			await runTests({
+				version: vscode_version,
+				extensionDevelopmentPath,
+				extensionTestsPath,
+				// note that passing environment variables seems to only work when invoking the tests via CLI
+				extensionTestsEnv: env,
+				// Disable any other extension
+				launchArgs: ["--disable-extensions", tmpWorkspace.name],
+			});
+		} finally {
+			try {
+				tmpWorkspace.removeCallback();
+			} catch (e) {
+				console.warn(`cleaning temporary directory has failed with error ${e}`);
+			}
 		}
 	}
 }
